+++ test/testgrid.jl
@@ -1075,46 +1075,6 @@ function createTest5BusNet(; cooldown = 0, hyst_pu = 0.0, qlim_min = nothing, ql
   return Bus5Net
 end
 
-function createTest3BusNet(; cooldown = 0, hyst_pu = 0.0, qlim_min = nothing, qlim_max = nothing)::Net
-  # Simple 3-bus network
-  #
-  #  ASTADT        STATION1
-  # <--|---------------|<--- Generator 
-  #    |-------       |
-  #            |      |
-  #            --------|<---- EXTERNALNETWORKINJECTION
-  #                 VERBUND  
-  Sbase_MVA = 100.0
-  netName = "test3bus"
-
-  r = 0.0
-  x = 0.4
-  s = 25.0
-  c_nf_per_km = 9.55
-  tanδ = 0.0
-
-  vm_pu_STATION1 = 1.027273
-  vm_pu_VERBUND = 1.018182
-
-  @debug "Creating $netName test network with qlim_min=$qlim_min, qlim_max=$qlim_max"
-
-  Bus3Net = Net(name = netName, baseMVA = Sbase_MVA, cooldown_iters = cooldown, q_hyst_pu = hyst_pu)
-
-  addBus!(net = Bus3Net, busName = "ASTADT", vn_kV = 110.0)
-  addBus!(net = Bus3Net, busName = "STATION1", vn_kV = 110.0)
-  addBus!(net = Bus3Net, busName = "VERBUND", vn_kV = 110.0)
-
-  addACLine!(net = Bus3Net, fromBus = "ASTADT", toBus = "STATION1", length = s, r = r, x = x, c_nf_per_km = c_nf_per_km, tanδ = tanδ)
-  addACLine!(net = Bus3Net, fromBus = "ASTADT", toBus = "VERBUND", length = s, r = r, x = x, c_nf_per_km = c_nf_per_km, tanδ = tanδ)
-  addACLine!(net = Bus3Net, fromBus = "VERBUND", toBus = "STATION1", length = s, r = r, x = x, c_nf_per_km = c_nf_per_km, tanδ = tanδ)
-
-  addProsumer!(net = Bus3Net, busName = "VERBUND", type = "EXTERNALNETWORKINJECTION", vm_pu = vm_pu_VERBUND, va_deg = 0.0, referencePri = "VERBUND")
-  addProsumer!(net = Bus3Net, busName = "STATION1", type = "SYNCHRONOUSMACHINE", p = 70.0, q = 33.2, vm_pu = vm_pu_STATION1, qMax = qlim_max, qMin = qlim_min)
-  addProsumer!(net = Bus3Net, busName = "ASTADT", type = "ENERGYCONSUMER", p = 100.0, q = 30.0)
-
-  return Bus3Net
-end
-
 function createTest2BusNet(; cooldown = 0, hyst_pu = 0.0, qlim_min = nothing, qlim_max = nothing)::Net
   # Simple 2-bus network
   #   Bus1            Bus2